refactor: centralize dm-crypt parameters in crypt.go with shared cryptsetup args
Mirror verity.go for the encryption side: move the EMWP dm-crypt
constants, GPT geometry, and key derivation into crypt.go, and give
packer and consumer a single CryptsetupArgs builder so their plain
dm-crypt mappings can never drift apart.
